I now do rendering through game objects instead of throgh a sprite sheet in the level... I got rid of collisions for the mean time until i know everything else is fine... i adjusted the way i increment frames... as well as adjusted value types for no problems for the computer...

+++ ShapesOfCombat/ShapesOfCombat/ShapesOfCombat/SOC_GameObject.h
@@ -13,7 +13,7 @@ class SOC_GameObject
 {
 public: 
 	SOC_GameObject();
-	SOC_GameObject(const wchar_t* filename, SOC_Graphics* newGraphics,int frameWidth, int frameHeight, int amountOfFrames);
+	SOC_GameObject(const wchar_t* filename, SOC_Graphics* newGraphics,float frameWidth, float frameHeight, int amountOfFrames);
 	~SOC_GameObject();
 
 	void Update();
