+++ ShapesOfCombat/ShapesOfCombat/ShapesOfCombat/SOC_Vector.h
@@ -23,7 +23,7 @@ struct SOC_Vector2
 		yVal = newYVal;
 	}
 
-	float magnitude() { return sqrt(xVal * xVal + yVal * yVal); }
+	float magnitude() { return (float)sqrt((double)xVal * (double)xVal + (double)yVal * (double)yVal); }
 
 	float Dot(SOC_Vector2 other)
 	{
